I hate filesystems almost as much as Fortran
Hg-commit: 9b9a3ced67554f8f2e0021fc71f35427d33051e8

+++ config/base.py
@@ -85,6 +85,8 @@ class Configure(script.Script):
     return
 
   def __del__(self):
+    if not os.path.isdir(self.tmpDir):
+      raise RuntimeError('How the hell could my tmpDir disappear?')
     shutil.rmtree(self.tmpDir)
     return
